requirements.txt
pdfminer.six>=20221105
python-docx>=0.8.11
nltk>=3.8.1
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0 